+++ CHANGELOG.md
@@ -79,12 +79,21 @@
   - fix color
   - add color
     - air-primary, air-primary-success, air-primary-alert, air-primary-copilot
-    - air-secondary, air-secondary-alert, air-secondary-accent, air-secondary-accent-1, air-secondary-accent-2, air-secondary-no-accent
-    - air-tertiary, air-tertiary-accent, air-tertiary-no-accent
+    - air-secondary, air-secondary-alert
   - @deprecate
     - default, danger, success, warning, primary, secondary, collab, ai
 * Accordion
   - fix color
+* Avatar,AvatarGroup
+  - fix color
+* Chip
+  - fix color
+  - add color
+    - air-primary, air-primary-success, air-primary-alert, air-primary-copilot
+    - air-secondary, air-secondary-alert, air-secondary-accent, air-secondary-accent-1, air-secondary-accent-2, air-secondary-no-accent
+    - air-tertiary, air-tertiary-accent, air-tertiary-no-accent
+  - @deprecate
+    - default, danger, success, warning, primary, secondary, collab, ai
 
 We change btn colors: